+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog for Django Glue
 
+## 0.4.2.2
+
+### Changes
+- Foreign Key fields now work with glue.
+- Move validation logic from Glue Request Handler to View.  
+
+
+
 ## 0.4.2.1
 
 ### Changes